backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.2.0
django-filter==23.5
drf-spectacular==0.27.1
django-cleanup==8.1.0
celery==5.3.6
redis==7.2.4
django-storages==1.14.2
boto3==1.34.34 